requirements.txt
dash==2.0.0
dash-ace==0.2.1 
dash-bootstrap-components==1.0.0
dash-core-components==2.0.0
dash-defer-js-import==0.0.2
dash-draggable==0.1.2
dash-html-components==2.0.0
dash-mantine-components==0.7.0
dash-pivottable==0.0.2
dash-table==5.0.0
dash-editor-components
Flask==2.1.0
geopandas==0.11.0
gunicorn==20.1.0
nltk==3.7
pandas==1.3.1
plotly==5.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyarrow==7.0.0
python-dateutil==2.8.1
python-dotenv==0.20.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sqlite==3.35.4
urllib3==1.26.6
wordcloud==1.8.1